+++ resources/person_related_resources/filmpersonrole_resource.py
@@ -1,4 +1,5 @@
 from flask import request, make_response, jsonify
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from flask_restplus import fields, Namespace, Resource
 from marshmallow import ValidationError
 
@@ -6,8 +7,8 @@ from models.models import FilmPersonRoleModel
 from schemas.schemas import FilmPersonRoleSchema
 
 # FilmPersonRole --------------------------------------------------------------
-model_name_singular = 'film_person_role'
-model_name_plural = 'films_persons_roles'
+model_name_singular = 'film-person-role'
+model_name_plural = 'films-persons-roles'
 model = FilmPersonRoleModel
 schema = FilmPersonRoleSchema()
 list_schema = FilmPersonRoleSchema(many=True)
@@ -28,7 +29,13 @@ class FilmPersonRoleResource(Resource):
             return schema.dump(model_data)
         return {'message': message_not_found}, 404
 
+    @jwt_required
     def delete(self, id):
+        claims = get_jwt_claims()
+
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+
         model_data = model.find_by_id(id)
         if model_data:
             model_data.delete_from_db()
@@ -36,8 +43,14 @@ class FilmPersonRoleResource(Resource):
                                f" Deleted successfully"}, 200
         return {'message': message_not_found}, 404
 
+    @jwt_required
     @namespace.expect(model_namespace)
     def put(self, id):
+        claims = get_jwt_claims()
+
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+
         model_data = model.find_by_id(id)
         model_json = request.get_json()
 
@@ -61,9 +74,15 @@ class FilmPersonRoleResourceList(Resource):
     def get(self):
         return list_schema.dump(model.find_all()), 200
 
+    @jwt_required
     @namespace.expect(model_namespace)
     @namespace.doc(f'Create an {model_name_singular}')
     def post(self):
+        claims = get_jwt_claims()
+
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+
         model_json = request.get_json()
         model_data = schema.load(model_json)
